src/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
jinja2
pypdf
google-genai
Python_dotenv
pymongo[srv]==3.12
datetime
python-docx
python-pptx
pandas
xlrd
openpyxl
ebooklib
pdf2image
bs4
pathlib
asyncio
motor